+++ Changelog.md
@@ -4,6 +4,8 @@
 
 ### Changes
 
+- ESDB: Added `transact({event, model, store, dispatch})` phase to the event processing flow. In this callback, you can call `dispatch` to generate and await sub-events, and calling ESModel will work too (any model can use the `dispatch` given via the constructor).
+  This requires the use of `AsyncLocalStorage`, and thus the minimum NodeJS version is now v12.17
 - SQLite: added `.runOnceOnOpen()` to register functions that should be run on the open database but shouldn't open the database
 - Renamed `dispatch()` to `addEvent()` inside the event processing flow. `dispatch()` still works but gives a deprecation warning. The `.dispatch()` method is not affected.